Add files via upload
Uploaded a Jupyter Notebook containing exercises and content related to data science tools, languages, and libraries. This notebook provides insights into the data science ecosystem and practical examples of Python programming for data analysis. It's intended to assist in learning and understanding essential concepts in the field of data science.